+++ src/main/java/ru/opensecreto/sigmacoin/vm/Memory.java
@@ -8,14 +8,14 @@ public interface Memory {
      * @param pointer position of word to get
      * @return word with specified pointer or 0x00 if word is unset
      */
-    public Word get(long pointer);
+    public Word get(int pointer);
 
     /**
      * Empty word at specified position ie. set 0x00 at specified poistion
      *
      * @param pointer position of word to empty
      */
-    public default void remove(long pointer) {
+    public default void remove(int pointer) {
         set(pointer, Word.WORD_0);
     }
 
@@ -25,6 +25,6 @@ public interface Memory {
      * @param pointer position of word to set
      * @param word    word to set at given position
      */
-    public void set(long pointer, Word word);
+    public void set(int pointer, Word word);
 
 }
